requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
pyyaml>=6.0
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.25.0
